refactor evaluation pipeline
