+++ tot/models/models_prophet.py
@@ -6,6 +6,7 @@ from typing import Type
 import pandas as pd
 
 from tot.df_utils import _check_min_df_len
+from tot.error_utils import raise_if
 from tot.models.models import Model
 from tot.models.utils import _get_seasons
 
@@ -46,6 +47,16 @@ class ProphetModel(Model):
             raise NotImplementedError(
                 "Quantiles for Prophet not supported in Test-of-Time. Remove interval_width from model input."
             )
+        model_params.pop("retrain", None)
+
+        norm_mode = model_params.pop("norm_mode", None)
+        norm_type = model_params.pop("norm_type", None)
+        norm_affine = model_params.pop("norm_affine", None)
+        raise_if(
+            norm_mode is not None or norm_type is not None or norm_affine is not None,
+            "Normalization layer not supported in darts models.",
+        )
+
         self.model = self.model_class(**model_params)
         if custom_seasonalities is not None:
             for seasonality in custom_seasonalities:
